msm: sde: Fix intf_sel for SDM845/SDM670/SDM660 targets
In 4.14 kernel the definition of the active interface for cont splash
was changed[1], which is invalid for older SoCs (SDM845/670/660), for them a
different bit[2] must be used to check the active interface. Hence add a
proper check for these legacy SoCs to the get_ctl_intf function and
protect it by checking the SDE HW revision.

[1] https://source.codeaurora.org/quic/la/kernel/msm-4.14/commit/?h=LA.UM.9.1.r1-11400-SMxxx0.0&id=75cc47bfa5f8ffb8ee43d6597ac2e8c74093cd91
[2] https://source.codeaurora.org/quic/la/kernel/msm-4.9/tree/drivers/gpu/drm/msm/sde/sde_rm.c?h=LA.UM.10.3.r1-00500-sdm845.0#n1433

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>